Rename module to oauth2ext
Long long ago, this started as an OIDC specific interface. Over time, it added
more features that are useful to OAuth2 as well - like the CLI flow, and Token
caching. Later, we moved it back to be a thinner wrapper over x/oauth2, and
added support for the JWT OAuth2 access token standard. With this in mind,
keeping it as an OIDC-specific library doesn't make sense.

Move the OIDC code into a sub-package, and name the module for a more OAuth2
expanded universe role. Re-name the middleware to note that it's OIDC specific,
in the future we might want to have a broader OAuth2 version.
