+++ claims/id_claims.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/lstoll/oidc/internal/th"
+	"github.com/lstoll/oauth2ext/internal/th"
 	"github.com/tink-crypto/tink-go/v2/jwt"
 )
