+++ clitoken/cache_darwin_cgo.go
@@ -19,8 +19,8 @@ import (
 	"syscall"
 	"unsafe"
 
-	"github.com/lstoll/oidc"
-	"github.com/lstoll/oidc/tokencache"
+	"github.com/lstoll/oauth2ext/oidc"
+	"github.com/lstoll/oauth2ext/tokencache"
 	"golang.org/x/oauth2"
 )
