+++ cmd/oidc-example-rp/main.go
@@ -6,7 +6,7 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/lstoll/oidc"
+	"github.com/lstoll/oauth2ext/oidc"
 	"golang.org/x/oauth2"
 )
