+++ cmd/oidccli/main.go
@@ -9,9 +9,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/lstoll/oidc"
-	"github.com/lstoll/oidc/clitoken"
-	"github.com/lstoll/oidc/tokencache"
+	"github.com/lstoll/oauth2ext/clitoken"
+	"github.com/lstoll/oauth2ext/oidc"
+	"github.com/lstoll/oauth2ext/tokencache"
 	"golang.org/x/oauth2"
 )
