+++ oidctest/provider.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/lstoll/oidc"
+	"github.com/lstoll/oauth2ext/oidc"
 	"github.com/tink-crypto/tink-go/v2/jwt"
 	"github.com/tink-crypto/tink-go/v2/keyset"
 )
